feat: move model to the `models` folder
