+++ keychain/api/field.py
@@ -14,7 +14,7 @@ from sqlalchemy.exc import IntegrityError
 
 from keychain import db
 from keychain.api.get_user_id import get_user_id
-from keychain.database.models import Field, Password
+from keychain.models import Field, Password
 
 logger = getLogger(__name__)
