+++ keychain/models/__init__.py
@@ -1,0 +1,4 @@
+from keychain.models.fields import Field
+from keychain.models.password import Password
+
+__all__ = ["Password", "Field"]
